+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Next]
+
+- Fix time synchronization to use local timezone offset (#98)
+
+
 ## [1.4.3] - 2025-06-22
 
 - Jupiter: Add MPPT sensors for PV voltage, current, and power (#82)